modified

+++ src/main/java/hexlet/code/controller/PostController.java
@@ -1,7 +1,9 @@
 package hexlet.code.controller;
 
+import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.model.Post;
 import hexlet.code.repository.PostRepository;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,7 +17,6 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/posts")
 public class PostController {
-    //private final List<Post> posts = new ArrayList<>();
     @Autowired
     private PostRepository postRepository;
 
@@ -33,38 +34,41 @@ public class PostController {
     @GetMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public Post showPost(@PathVariable Long id) {
+        var post = postRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Post with id " + id + " not found"));
 
-        return postRepository.findById(id).get();
-        //var post = posts.stream()
-        //        .filter(p -> p.getSlug().equals(id))
-        //        .findFirst();
-        //return ResponseEntity.of(post);
+        return post;
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public Post createPost(@RequestBody Post post) {
-
-        return postRepository.save(post);
+    public ResponseEntity<Post> createPost(@Valid @RequestBody Post post) {
+        var savedPost = postRepository.save(post);
+        return ResponseEntity
+                .created(URI.create("/api/posts/" + savedPost.getId()))
+                .body(savedPost);
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public Post updatePost(@PathVariable Long id, @RequestBody Post data) {
-        var maybePost = postRepository.findById(id);
-        if (maybePost.isPresent()) {
-            var post = maybePost.get();
-            post.setTitle(data.getTitle());
-            post.setContent(data.getContent());
-            post.setPublished(data.isPublished());
-            return postRepository.save(post);
-        }
-        return null;
+    public Post updatePost(@PathVariable Long id, @Valid @RequestBody Post data) {
+        var post = postRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Post with id " + id + " not found"));
+
+        post.setTitle(data.getTitle());
+        post.setContent(data.getContent());
+        post.setPublished(data.isPublished());
+
+        return postRepository.save(post);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void destroyPost(@PathVariable Long id) {
+        var post = postRepository.findById(id);
+        if (post.isEmpty()) {
+            throw new ResourceNotFoundException("Post with ID " + id + " not found");
+        }
         postRepository.deleteById(id);
     }
 }
