+++ src/main/java/hexlet/code/controller/UserController.java
@@ -1,7 +1,9 @@
 package hexlet.code.controller;
 
+import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.model.User;
 import hexlet.code.repository.UserRepository;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,7 +17,6 @@ import java.util.concurrent.atomic.AtomicInteger;
 @RestController
 @RequestMapping("/api/users")
 public class UserController {
-    //private final List<User> users = new ArrayList<>();
     //генерируем уникальный id
     //private final AtomicInteger userIdCounter = new AtomicInteger(1);
     @Autowired
@@ -35,37 +36,44 @@ public class UserController {
     @GetMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public User showUser(@PathVariable Long id) {
-        return userRepository.findById(id).get();
+        var user = userRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
+        return user;
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public User createUser(@RequestBody User user) {
+    public ResponseEntity<User> createUser(@Valid @RequestBody User user) {
         //user.setId((long) userIdCounter.getAndIncrement());
         //users.add(user);
         //URI location = URI.create("/api/users/" + user.getId());
-        return userRepository.save(user);
+        var savedUser = userRepository.save(user);
+        return ResponseEntity
+                .created(URI.create("/api/users/" + savedUser.getId()))
+                .body(savedUser);
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public User updateUser(@PathVariable Long id, @RequestBody User data) {
-        var maybeUser = userRepository.findById(id);
-        if(maybeUser.isPresent()) {
-            var user = maybeUser.get();
-            user.setEmail(data.getEmail());
-            user.setFirstName(data.getFirstName());
-            user.setLastName(data.getLastName());
-            user.setBirthday(data.getBirthday());
+    public User updateUser(@PathVariable Long id, @Valid @RequestBody User data) {
+        var user = userRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
 
-            return userRepository.save(user);
-        }
-        return null;
+        user.setEmail(data.getEmail());
+        user.setFirstName(data.getFirstName());
+        user.setLastName(data.getLastName());
+        user.setBirthday(data.getBirthday());
+
+        return userRepository.save(user);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void destroyUser(@PathVariable long id) {
+    public void destroyUser(@PathVariable Long id) {
+        var user = userRepository.findById(id);
+        if (user.isEmpty()) {
+            throw new ResourceNotFoundException("User with ID " + id + " not found");
+        }
         userRepository.deleteById(id);
     }
 }
